--- src/components/Nav/Nav.js
@@ -1,5 +1,0 @@
-function Nav() {
-  return <div>나 응애나브</div>;
-}
-
-export default Nav;
