+++ src/components/Nav/Nav.scss
@@ -1,0 +1,136 @@
+@mixin flex-center($direction: center) {
+  display: flex;
+  justify-content: $direction;
+  align-items: center;
+}
+
+.navClass {
+  position: sticky;
+  top: 0;
+  z-index: 1000;
+  /* nav-bar */
+  .nav {
+    @include flex-center;
+    width: 100%;
+    height: 60px;
+    background-color: white;
+    border-bottom: 1px solid rgba(0, 0, 0, 0.2);
+  }
+
+  .nav-container {
+    @include flex-center($direction: space-between);
+    width: 975px;
+    padding: 0 20px;
+  }
+
+  .nav-container__logo {
+    @include flex-center;
+    gap: 1rem;
+    margin-right: 40px;
+  }
+
+  #instagram {
+    margin-top: 5px;
+    font-size: 34px;
+  }
+
+  .nav-logo {
+    font-family: var(--logo-font);
+    font-size: 30px;
+  }
+
+  .nav-logo__separate {
+    width: 1px;
+    height: 30px;
+    background-color: black;
+  }
+
+  .nav-container__search {
+    @include flex-center;
+    width: 280px;
+    height: fit-content;
+  }
+
+  .search-input__input {
+    width: 260px;
+    height: 35px;
+    border: none;
+    border-radius: 0.6rem;
+    background-color: rgba(157, 151, 151, 0.1);
+    padding: 0.8rem;
+
+    &[placeholder] {
+      font-size: 15px;
+      font-weight: 300;
+      padding-left: 45px;
+    }
+
+    &:focus {
+      outline: none;
+    }
+  }
+
+  .search-input {
+    position: relative;
+  }
+
+  #search {
+    position: absolute;
+    top: 8px;
+    left: 15px;
+    font-size: 20px;
+    color: rgba(0, 0, 0, 0.3);
+  }
+
+  #exit {
+    display: none;
+    position: absolute;
+    font-size: 16px;
+    top: 9px;
+    right: 15px;
+    color: rgba(0, 0, 0, 0.3);
+  }
+
+  .nav-container__icon {
+    @include flex-center($direction: space-between);
+    width: auto;
+    gap: 0.9rem;
+  }
+
+  .nav-icon {
+    @include flex-center;
+    width: 30px;
+    height: 30px;
+
+    a {
+      i {
+        font-size: 28px;
+      }
+    }
+  }
+
+  .nav-profile__image {
+    width: 30px;
+    height: 30px;
+    border-radius: 50%;
+
+    a {
+      img {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+        border-radius: 50%;
+      }
+    }
+  }
+
+  /* javascript */
+
+  .hidden {
+    display: none;
+  }
+
+  .search-input__input[placeholder].focus {
+    padding-left: 25px;
+  }
+}
