+++ src/components/comment/comment.scss
@@ -1,0 +1,75 @@
+@mixin flex-center($direction: center) {
+  display: flex;
+  justify-content: $direction;
+  align-items: center;
+}
+
+.description-comment {
+  @include flex-center($direction: space-between);
+  margin-top: 3px;
+  margin-bottom: 3px;
+}
+
+.comment-left {
+  @include flex-center($direction: flex-start);
+  gap: 0.3rem;
+}
+
+.user-name {
+  font-weight: 600;
+}
+
+.comment-right {
+  @include flex-center($direction: space-between);
+  gap: 0.3rem;
+
+  i {
+    color: rgba(0, 0, 0, 0.5);
+
+    &:hover {
+      cursor: pointer;
+      color: #000;
+    }
+  }
+}
+
+/* commment submit */
+
+.comment-submit {
+  @include flex-center($direction: space-between);
+  width: 100%;
+  height: 50px;
+  padding: 10px 12px;
+  padding-left: 6px;
+  border-top: 1px solid rgba(0, 0, 0, 0.1);
+}
+
+.smile-icon {
+  width: 40px;
+  height: 40px;
+  background-image: url(../../assets/images/wonikChoi/instargram-smile.webp);
+  background-size: cover;
+  background-repeat: none;
+}
+
+.comment-input {
+  width: 375px;
+  height: 30px;
+  border: none;
+
+  &:focus {
+    outline: none;
+  }
+
+  &[placeholder] {
+    font-size: 13px;
+  }
+}
+
+.comment-btn {
+  border: none;
+  background-color: white;
+  color: #0597f2;
+  opacity: 0.5;
+  font-weight: 600;
+}
