+++ src/components/feed/feed.scss
@@ -1,0 +1,131 @@
+@mixin flex-center($direction: center) {
+  display: flex;
+  justify-content: $direction;
+  align-items: center;
+}
+
+.feed {
+  /* main article */
+  .main-article {
+    width: 100%;
+    height: auto;
+    background-color: white;
+    margin: 15px 0;
+    border: 1px solid rgba(0, 0, 0, 0.2);
+    border-radius: 0.4rem;
+  }
+
+  .main-article__nav {
+    @include flex-center($direction: space-between);
+    width: 100%;
+    height: 60px;
+    padding: 10px;
+  }
+
+  .nav-artivle__container {
+    @include flex-center;
+    gap: 0.5rem;
+    width: auto;
+    height: auto;
+    padding: 0px;
+  }
+
+  .nav-image {
+    width: 46px;
+    height: 46px;
+    border: 2px solid orange;
+    border-radius: 50%;
+
+    img {
+      width: 42px;
+      height: 42px;
+      border-radius: inherit;
+      object-fit: cover;
+      object-position: center;
+      padding: 0.1rem;
+    }
+  }
+
+  .nav-title {
+    font-size: 15px;
+    font-weight: 700;
+    line-height: 23px;
+  }
+
+  .nav-subtitle {
+    font-size: 13.5px;
+    color: rgba(0, 0, 0, 0.9);
+  }
+
+  .nav-icon {
+    i {
+      font-size: 20px;
+    }
+  }
+
+  /* picture */
+
+  .main-article__picture {
+    width: 100%;
+    height: 470px;
+
+    img {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+    }
+  }
+
+  /* heart */
+
+  .main-article__heart {
+    @include flex-center($direction: space-between);
+    width: 100%;
+    height: 50px;
+    padding: 10px 4px;
+    margin-bottom: 5px;
+  }
+
+  .heart-container {
+    @include flex-center($direction: flex-start);
+    width: 100%;
+    height: auto;
+  }
+
+  .heart-icon {
+    @include flex-center;
+    width: 40px;
+    height: 40px;
+
+    i {
+      font-size: 27px;
+    }
+  }
+
+  /* like */
+
+  .main-article__like {
+    width: 100%;
+    height: auto;
+    padding: 0 12px;
+    margin-bottom: 10px;
+  }
+
+  .like-span {
+    font-size: 14px;
+    font-weight: 600;
+  }
+
+  /* decription */
+
+  .main-article__description {
+    width: 100%;
+    height: auto;
+    padding: 0 12px;
+    font-size: 15px;
+  }
+
+  .userId {
+    font-weight: 600;
+  }
+}
