+++ src/components/recommendBox/recommendBox.scss
@@ -1,0 +1,61 @@
+@mixin flex-center($direction: center) {
+  display: flex;
+  justify-content: $direction;
+  align-items: center;
+}
+
+.friendBox {
+  .friends-container {
+    width: 100%;
+    height: auto;
+    margin-top: 15px;
+  }
+
+  .friends {
+    @include flex-center($direction: space-between);
+    width: 100%;
+    height: 52px;
+    padding: 5px;
+  }
+
+  .friends-left {
+    @include flex-center($direction: space-between);
+    gap: 0.7rem;
+  }
+
+  .friends-image {
+    width: 30px;
+    height: 30px;
+    border-radius: 50%;
+
+    img {
+      width: 100%;
+      height: 100%;
+      border-radius: inherit;
+      object-fit: cover;
+      object-position: center;
+    }
+  }
+
+  .friends-name {
+    display: flex;
+    flex-direction: column;
+    gap: 0.4rem;
+  }
+
+  .friends-user {
+    font-size: 14px;
+    font-weight: 700;
+  }
+
+  .friends-sub {
+    font-size: 12px;
+    color: rgba(0, 0, 0, 0.5);
+  }
+
+  .follow {
+    font-size: 12px;
+    color: #0597f2;
+    font-weight: 600;
+  }
+}
